+++ src/core.jl
@@ -1,7 +1,26 @@
+codegen_ast_func(; kw...) = codegen_ast(JLFunction(; kw...))
+
+struct JLCall
+    func
+    args::Vector
+    kwargs::Union{Nothing,Vector}
+end
+
+JLCall(; func, args, kwargs) = JLCall(func, args, kwargs)
+
+function codegen_ast_call(x::JLCall)
+    call = Expr(:call, x.func)
+    isnothing(x.kwargs) || push!(call.args, Expr(:parameters, x.kwargs...))
+    append!(call.args, x.args)
+    return call
+end
+
+codegen_ast_call(; kw...) = codegen_ast_call(JLCall(; kw...))
+
 function generate_permuted_methods(args; perms=nothing, kw...)
     perms = something(perms, permutations(1:length(args)))
     return map(perms) do p
-        JLFunction(; args=args[p], kw...) |> codegen_ast
+        codegen_ast_func(; args=args[p], kw...)
     end
 end
 
@@ -25,11 +44,11 @@ function generate_permutable_struct(expr)
     # Generate outer constructors for all permutations
     perms = collect(permutations(1:length(fields)))[2:end]
     constructors = map(perms) do p
-        JLFunction(; name, args=fields[p], body) |> codegen_ast
+        codegen_ast_func(; name, args=fields[p], body)
     end
 
     # Generate keyword constructor
-    kw_constructor = JLFunction(; name, kwargs=fields, body) |> codegen_ast
+    kw_constructor = codegen_ast_func(; name, kwargs=fields, body)
 
     return Expr(:block, expr, constructors..., kw_constructor)
 end